1.4 Installation & Setup
Relevant Packages for Unity Multiplayer:
 1) Unity Netcode
 2) Unity Multiplay
 3) Unity Matchmaker

For a basic multiplayer game to start, you need a network manager. The network manager takes care of the hosting, clientele, and servers. Additionally, you can add a player prefab object into the network manager, and this will be instantiated each time a player joins the game. There are also network prefabs that can be added to an array. Finally, the network manager uses a Unity Transform component to send and receive information from the Unity server back to the player machines.

For the player prefab, each player needs two components: a network object and a network transform.

The network object is responsible for creating a unique network ID for the object, so that the player is properly assigned to the game. It also provides client ownership over the instantiated object. When the client joins/leaves the game, this component communicates with the network manager to tell all other clients to instantiate/destroy the object.

The network transform component is responsible for showing the transform properties of the object to other clients. You can enable/disable properties that you don't want other clients to see, like the x, y, or z position, and other values like the x, y, or z scale.

Finally, to host the game and join the server, you need to use the network manager. First, a game must be hosted. For now, this can be done by running the game in Unity editor, opening the network manager, and clicking "start host". To get another player in the game, you need to build and run the game and apply simple logic (see join server script) to join the game as a client.