2.2 Importing Assets
GameDev assets were added in this course lecture. I will follow along with the given assets for ease of use.

The player tank is set up with the bottom and top parts, named treads and turret pivot, respectively. The parent player game object is only responsible for updating the x and y position in the game, and the child objects that are rendered will be responsible for the other transform values.

For the treads, there is a 2D box collider, so this will be responsible for collision logic. This will send its z rotation to the server so people can see the tank's movement rotation. The tank body is also a child of the treads object, and that just has a sprite renderer.

The turret pivot is an empty object and is responsible for the z position and z rotation for itself and child game objects. The turret sprite renderer is a child of the object.