2.17 Compat Polish
Small difference in setup between instructor and my version. SpendCoins() method returns a bool for whether coins were spent or not, and proper logic is handled in the Coin Wallet script instead of the Player Shooter script.

Additionally, a coroutine was used to set a bool to true/false to handle new shooting logic given the fire rate.