3.5 Auth Improvements
This lecture covered exception handling to ensure our authentication process runs smoothly. Two exceptions were handled:

1) authentication error. This occurs if there is something wrong on the client's end. This may be due to invalid or expired player ID, corrupted identity data, or authenticating multiple times. This error is on the user's part

2) request failed exception. This exception is for more networking problems, connection timeouts, and invalid project configurations. Examples include no internet connection, Unity servers are down, invalid project ID, and other networking errors

Additionally, the anonymous sign in process performs one last check if the player is still authenticating at the end of the method. If this runs, that means the player unsuccessfully authenticated due to a timeout (no more attmepts).

Finally, before attempting authentication, we check if the player is currently authenticating or is not yet authenticated. If so, we wait for the auth state to change to any other auth state before continuing.