3.7 Allocating a Relay
Key takeaways:

1) What exactly is a relay? It is a server that players can connect to. Instead of the dedicated server being on the player's machine, we connect to Unity's relay services to handle the connections for us.

2) What is an allocation? It is a "virtual room" for the host and clients. This holds information about each session, and we can specify the number of connections allowed for each allocation.

3) What is a transport? It is somewhat self-explanatory, where data is transported from one client to another.

4) What is relay server data? It contains information relevant for players connecting to a relay server. Typically, games use "udp" for fast connections.

With that in mind, the host game manager has a new method for starting the game as a host. This attempts to create an allocation. If so, we try to receive a join code from the allocation, passing in the newly created allocation ID. This join code can be passed to others to join a game.

After the allocation setup, we set up the transport. We store reference to the unity transport component on our network manager, create a new relay server data object, passing in our allocation and the "udp" connection type. Finally, we set up the relay server data and tell the network manager to start the player as a host (this is the equivelant of pressing "start host" in the network manager game object). Last, we change the scene to the game scene.

This setup can be used as a template for creating hosts in future multiplayer objects. Do not get too caught up on the connection logic.