3.8 Joining A Relay
The setup for a client joining the game is extremely similar to the setup for starting as a host. The client game manage has a method for joining as a client, passing the join code to connect to the host. We use the join code to initialize our allocation. We set up the unity transport on the network manager, and we call the StartClient method to join the relay with the host.

My setup for the Unity relay was incorrect in the last commit. I did not create the relay on the Unity cloud. This commit and future commits will work with that now.

Note: the client will automatically be placed in the network scene when joining the server, so they do not need to handle scene transition logic.