3.15 Networking Improvements
The title for this is misleading. This section focuses on editor only coding. In this game, we have the editor always start on the first scene in the game, regardless of our current scene. We need to put editor scripts in a folder under assets called "Editor." This folder will not be used outside of the Unity editor application.